block: set op to REQ_OP
This patch converts the request related block layer code to set
request->op to a REQ_OP and cmd_flags to rq_flag_bits.

There is some tmp compat code when setting up cmd_flags so it
still carries both the op and flags. It will be removed in
in later patches in this set when I have converted all drivers.

I have not been able to test the mq paths with real mq hardware.

v4:
1. Fix missing docbook comment for get_request.

v3:
1. Removed __get_request cmd_flags compat code.

Signed-off-by: Mike Christie <mchristi@redhat.com>