RAID 4/5/6: Fine-grained TRIM enable for ZDM
Currently RAID 4/5/6 requires an module parameter to enable
discard bios.

This was done due the to general inadequacies of relying on
hardware to correctly self identify.

Since ZDM does not rely on lower level device support to for
TRIM handling it seems reasonable to allow another method to
communicate to MD-RAID that enabling discard is safe.

This method is also more targeted. Devices are not expected
to set this parameter directly. And the kernel module parameter
which assumes devices are unsafe remains intact.

Signed-off-by: Shaun Tancheff <shaun.tancheff@seagate.com>