Losing bits on request.cmd_flags
In a few places a temporary value smaller than a cmd_flags
is used to test for bits and or build up a new cmd_flags.

Change to use explict u64 values where appropriate.

This patch is place holder for: Mike Christie's separate operations ... series
https://lkml.kernel.org/r/1465155145-10812-1-git-send-email-mchristi@redhat.com

Once Mike's patches are stablized this patch can be dropped.

Signed-off-by: Shaun Tancheff <shaun.tancheff@seagate.com>