Add bio/request flags for using ZBC/ZAC commands
T10 ZBC and T13 ZAC specify operations for Zoned devices.

To be able to access the zone information and open and close zones
adding op's for:
  - Report zones command: REQ_OP_ZONE_REPORT
  - Open zone: REQ_OP_ZONE_OPEN
  - Close zone: REQ_OP_ZONE_CLOSE
  - Reset Write Pointer: REQ_OP_ZONE_RESET
to be used to create struct bio / struct request to issue
ZBC commands.

Report zones is a device read that requires a buffer.
Open, Close and Reset are device commands that have no associated
data transfer.

The Finish zone command is intentionally not implimented as there is no
current use case for that operation.

Report zones currently defaults to reporting on all zones. It expected
that support for the zone option flag will piggy back on streamid
support. The report option flag is useful as it can reduce the number
of zones in each report, but not critical.

Signed-off-by: Shaun Tancheff <shaun.tancheff@seagate.com>
---
v4:
 - Rebase on linux-next tag next-20160615.
 - Change bio flags to bio op's
V3:
 - Rebase on Mike Cristie's separate bio operations
 - Update blkzoned_api.h to include report zones PARTIAL bit.

V2:
 - Changed bi_rw to op_flags clarify sepeartion of bio op from flags.
 - Fixed memory leak in blkdev_issue_zone_report failing to put_bio().
 - Documented opt in blkdev_issue_zone_report.
 - Removed include/uapi/linux/fs.h from this patch.