Add bio/request flags to issue ZBC/ZAC commands
Add op flags to access to zone information as well as open, close
and reset zones:
  - Report zones command: REQ_OP_ZONE_REPORT
  - Open zone: REQ_OP_ZONE_OPEN
  - Close zone: REQ_OP_ZONE_CLOSE
  - Reset Write Pointer: REQ_OP_ZONE_RESET

These op flags can be used to create bio's to control zoned devices
through the block layer.

This is useful for filesystems and device mappers that need explicit
control of zoned devices such as Host Mananged and Host Aware SMR drives,

Report zones is a device read that requires a buffer.
Open, Close and Reset are device commands that have no associated
data transfer.

The Finish zone command is intentionally not implimented as there is no
current use case for that operation.

Report zones currently defaults to reporting on all zones. It expected
that support for the zone option flag will piggy back on streamid
support. The report option flag is useful as it can reduce the number
of zones in each report, but not critical.

Signed-off-by: Shaun Tancheff <shaun.tancheff@seagate.com>