Add support for URSWRZ zoned configuration bit
Currently devices are presumed to follow unrestricted reads in that
read failures indicated bad media or partial writes.

Host Managed devices can optional enable a restricted read mode that
returns errors for reads beyond the current zone write pointer or
reads the cross zone boundaries.

This restricted read mode is indicated in the URSWRZ bit of the zoned
capability.

This patch exposes the URSWRZ flag via sysfs.

A proper default (1) is reported for devices that do not support
restricted reads.

Signed-off-by: Shaun Tancheff <shaun.tancheff@seagate.com>