fsnotify: drop notification_mutex before destroying event
fsnotify_flush_notify() and fanotify_release() destroy notification event
while holding notification_mutex.  Destruction of fanotify event includes
a path_put() call which may end up calling into a filesystem to delete an
inode if we happen to be the last holders of dentry reference which
happens to be the last holder of inode reference.  That may violate lock
ordering for some filesystems since notification_mutex is also acquired e.
g.  during write when generating fanotify event.  Also this is the only
thing that forces notification_mutex to be a sleeping lock.  So drop
notification_mutex before destroying a notification event.

Link: http://lkml.kernel.org/r/1473797711-14111-4-git-send-email-jack@suse.cz
Signed-off-by: Jan Kara <jack@suse.cz>
Cc: Miklos Szeredi <mszeredi@redhat.com>
Cc: Lino Sanfilippo <LinoSanfilippo@gmx.de>
Cc: Eric Paris <eparis@redhat.com>
Cc: Al Viro <viro@ZenIV.linux.org.uk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>