ext4: feature pre-allocated folios for buffered i/o
This reduces the page allocation contention for buffered i/o
by grabbing a pool of pages suitable for being added to the
page cache.

This gives about 5% better throughput for large writes that
need new pages to be added to the page cache.

A new sysfs entry prealloc_folios is used to enable and limit
the number of pre-allocated pages created during I/O

Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>